Isolate attribute methods from development dependencies
Issues seeding the db on the remote server due to dev deps being
imported in various files that are used by seed.py

This isolates the deps in seed.py from other dev deps.
